Refactor invoice tests to use decimal type for amounts and update InvoiceFaker to generate decimal amounts for consistency

+++ tests/AppDomain.Tests/Integration/Invoices/CancelInvoiceIntegrationTests.cs
@@ -1,5 +1,6 @@
 // Copyright (c) OrgName. All rights reserved.
 
+using AppDomain.Common.Grpc;
 using AppDomain.Invoices.Grpc;
 using AppDomain.Tests.Integration._Internal;
 using AppDomain.Tests.Integration._Internal.TestDataGenerators;
@@ -21,7 +22,7 @@ public class CancelInvoiceIntegrationTests(IntegrationTestFixture fixture) : Int
 
         // Arrange - Create an invoice first
         var createRequest = _invoiceFaker
-            .WithAmount(100.00)
+            .WithAmount(100.00m)
             .WithCurrency("USD")
             .Generate();
 
@@ -39,7 +40,7 @@ public class CancelInvoiceIntegrationTests(IntegrationTestFixture fixture) : Int
         cancelledInvoice.InvoiceId.ShouldBe(createdInvoice.InvoiceId);
         cancelledInvoice.Status.ShouldBe("Cancelled");
         cancelledInvoice.Name.ShouldBe(createRequest.Name);
-        cancelledInvoice.Amount.ShouldBe(100.00);
+        ((decimal)cancelledInvoice.Amount).ShouldBe(100.00m);
 
         // Verify in database
         var dbInvoice = await connection.QuerySingleOrDefaultAsync(
@@ -91,7 +92,7 @@ public class CancelInvoiceIntegrationTests(IntegrationTestFixture fixture) : Int
 
         // Arrange - Create and cancel an invoice
         var createRequest = _invoiceFaker
-            .WithAmount(100.00)
+            .WithAmount(100.00m)
             .WithCurrency("USD")
             .Generate();
